note_taking_app/requirements.txt
Flask
SpeechRecognition
mysql-connector-python
python-dotenv
fpdf
python-docx